==> notebooks/1.0-bk-heremaps-data-extraction.ipynb <==
# Extracting Data from HereMaps (here.com)
<br>

_Author: Bala Krishnamoorthy_

#### Goal

Access here.com's api to extract information on road closures, incidents, etc.

To be completed...

### Data Gathering

# Import Libraries

import json, requests
import time, datetime
import pandas as pd

# Viewing configs for notebook
%matplotlib inline
%config InlineBackend.print_figure_kwargs={'facecolor' : 'w'}
# Code below allows multiple console outputs to be generated without print statements
from IPython.core.interactiveshell import InteractiveShell 
InteractiveShell.ast_node_interactivity = "all"

# Imports my API credentials stored locally. These are **not** publically available. To use
# the API access code below, please create your own access credentials on here.com and execute
# the code below. 
import config 

# Specify Bounding Box Coordinates (latitude, longitude)

# Bounding Box Coordinates - test case for larger area
# west_south_east_north = '-122.384674,47.515083,-122.118271,47.644727' # Seattle + Bellevue 
city = 'Houston, TX' # Using Twitter's assumption for bounding box around Houston
top_left = '30.1546646,-95.823268'
bottom_right = '29.522325,-95.069705'

url = 'https://traffic.api.here.com/traffic/6.3/\
incidents.json?app_id=' + config.app_id + '&app_code=' + config.app_code + '&bbox=' + \
top_left + ';' + bottom_right + '&criticality=critical'
url;

# Criticality = critical indicates that only road closures are requested from the 
# HereMaps API.

res = requests.get(url)
date_requested = time.strftime('%Y-%m-%d')
time_requested = time.strftime('%H:%M') # 24-hour format
date_requested; time_requested 

res.status_code

hmaps_json = res.json()
hmaps_json;

# Inspect json file
# hmaps_json.keys()
# hmaps_json['TRAFFIC_ITEMS'].keys()
# hmaps_json['diagnostic'].keys()
hmaps_json['TRAFFIC_ITEMS']['TRAFFIC_ITEM'];

incident_dict = hmaps_json['TRAFFIC_ITEMS']['TRAFFIC_ITEM']
incident_dict[0].keys()

incident_df = pd.DataFrame(incident_dict)
incident_df.head()

# Number of total incidents in specified bounding box
num_incidents = incident_df.shape[0]
num_incidents

incident_df_2 = incident_df.drop(axis=1, columns=['ABBREVIATION', 'ORIGINAL_TRAFFIC_ITEM_ID', 'PRODUCT', 
                                  'TRAFFIC_ITEM_ID', 'mid'])
incident_df_2.rename(str.lower, axis=1, inplace=True)
incident_df_2.head()

incident_df_2.columns

loc_all = pd.DataFrame(list(incident_df_2['location']))
loc_all.rename(str.lower, axis=1, inplace=True)
loc_all.head()

# Fill NaN Values with empty dictionaries so that dataframes can continue to be created
loc_all = loc_all.applymap(lambda x: {} if pd.isnull(x) else x)
loc_all.head()

loc_region = pd.DataFrame(list(loc_all['political_boundary']))
loc_region.rename(str.lower, axis=1, inplace=True)
loc_region.head()

loc_intersec = pd.DataFrame(list(loc_all['intersection']))
loc_intersec.rename(str.lower, axis=1, inplace=True)
loc_intersec.head()

loc_coord = pd.DataFrame(list(loc_all['geoloc']))
loc_coord.rename(str.lower, axis=1, inplace=True)
loc_coord.head()

### Outputs

**Potential Additional Updates**:
- Additional columns:
    - Type of incident (planned, unplanned (blocked by tree, power failure, etc.))

# Instantiate an empty dataframe that captures only the columns of interest
columns = ['city', 'source', 'criticality', 'intersection', 'start_lat_long', 
           'end_lat_long', 'start_time', 'end_time', 'entry_time', 'bbox_top_left', 
           'bbox_bottom_right', 'date_requested', 'time_requested']
output_df = pd.DataFrame(index=incident_df_2.index, columns=columns)
output_df.fillna('', inplace=True)
output_df.loc[:, 'city'] = city

output_df.head()

**Populate the Outputs dataframe using data from HereMaps**

for index in incident_df_2.index:
    # Incident-Agnostic Parameters
    output_df['bbox_top_left'][index] = str(tuple(top_left.split(',')))
    output_df['bbox_bottom_right'][index] = str(tuple(bottom_right.split(',')))
    output_df['date_requested'][index] = date_requested
    output_df['time_requested'][index] = time_requested
    output_df['source'][index] = 'Here Maps API'
    # Incident-Specific Parameters
    if pd.isna(incident_df_2['criticality'][index]) == False:
        output_df['criticality'][index] = incident_df_2['criticality'][index]['DESCRIPTION']
    else:
        output_df['criticality'][index] = '-'
    if pd.isna(loc_coord['origin'][index]) == False:
        output_df['start_lat_long'][index] = tuple([loc_coord['origin'][index]['LATITUDE'],
                                   loc_coord['origin'][index]['LONGITUDE']])
    else: 
        output_df['start_lat_long'][index] = '-'
    if pd.isna(loc_coord['to'][index][0]) == False:
        output_df['end_lat_long'][index] = tuple([loc_coord['to'][index][0]['LATITUDE'],
                                 loc_coord['to'][index][0]['LONGITUDE']])
    else:
        output_df['end_lat_long'][index] = '-'
    output_df['intersection'][index] = loc_all['intersection'][index]
    output_df['start_time'][index] = incident_df_2['start_time'][index]
    output_df['end_time'][index] = incident_df_2['end_time'][index]
    output_df['entry_time'][index] = incident_df_2['entry_time'][index]                           

# Final Output DataFrame
output_df

# Write Output DataFrame to csv file
output_df.to_csv('../data/2-interim/here_maps_output_' + \
                 city + '_' + time.strftime('%Y-%m-%d-%I%p'), index=False)
