==> notebooks/1.0-bk-tweets-location-extraction.ipynb <==
# Extracting location information from tweets <br>

_Author: Bala Krishnamoorthy_

import pandas as pd
import re

tweets_df = pd.read_csv('../data/tweets_with_intersection.csv');

tweets_df.drop(axis=1, columns='Unnamed: 0', inplace=True)

tweets_df.head()

tweets_df_2 = tweets_df[['text', 'place.bounding_box.coordinates', 'place.full_name']]
tweets_df_2.head()

# Number of tweets by location 
tweets_df['place.full_name'].value_counts()[:5]

Fort Worth has the most number of tweets about road blockages. Examine tweets from this city. 

# Fort Worth, TX Tweets as a Series
tweets_text_fw = tweets_df[tweets_df['place.full_name'] == 'Fort Worth, TX'][['text']]
tweets_text_fw.reset_index(inplace=True, drop=True)
tweets_text_fw.head()
for i in tweets_text_fw['text']:
    print(i)

Part 1 - Filter tweets to only use data on closed roads:

# Create new column
tweets_df['closed_road_check'] = ''
# Enter 1 if tweet indicates 'closed road', 0 if not
closed_road_indicators = ['Closed due to', 'Road closed']
index = 0
for i in tweets_df['text']:
    if any(x in i for x in closed_road_indicators):
        tweets_df['closed_road_check'][index] = 1
    else:
        tweets_df['closed_road_check'][index] = 0

tweets_df['closed_road_check']

### Filtered Tweets Below (Only Closed Roads)

tweets_classed_df = pd.read_csv('../data/2-interim/twitter_corpus.csv', sep='\t', 
                                encoding='latin-1')
tweets_classed_df.drop(axis=1, columns='tweed_id', inplace=True)

tweets_classed_df.head()

# 2 = closed
final_tweets_df = tweets_classed_df[tweets_classed_df['rating'] == 2]
final_tweets_df.head()

# Combine with original df (join by index) to pull city info
final_tweets_df = pd.merge(final_tweets_df, tweets_df, left_index=True, 
                           right_index=True)

final_tweets_df = final_tweets_df[['tweet','created_at',
                                   'place.bounding_box.coordinates','place.full_name']]
final_tweets_df.reset_index(drop=True, inplace=True)

final_tweets_df.head()

final_tweets_df['place.full_name'].value_counts()

Regex Step 2 - Extract street information:

2.1: Isolate street names and intersections

#### Choose City in Code Block Below

city = 'Houston, TX'
city_tweets = final_tweets_df[final_tweets_df['place.full_name'] == city]['tweet']
city_tweets.reset_index(drop=True, inplace=True)
intersection_list = []
bbox

# Extract bounding box (i.e. bbox) for chosen city -- Needed to extract HereMaps data for same
# city
bbox_raw = eval(final_tweets_df[final_tweets_df['place.full_name'] == city]\
['place.bounding_box.coordinates'][6])[0]

# twitter bbox is always defined by 4 corners of the bbox
bbox_top_left = str(bbox_raw[3][1]) + ',' + str(bbox_raw[3][0])
bbox_bottom_right = str(bbox_raw[1][1]) + ',' + str(bbox_raw[1][0])
print('bbox_top_left (lat,long):', bbox_top_left)
print('bbox_bottom_right (lat,long):', bbox_bottom_right)

# Generate string of "useful" tweets
for i in range(city_tweets.shape[0]):
    # Regex Pattern 1
    intersection_list.append(re.findall(' on (.+)[#]', city_tweets[i]))
    if i == (city_tweets.shape[0] - 1):
        print('intersection list:', intersection_list)
        print()
        print('number of tweets available:', city_tweets.shape[0])
        print('number of patterns found:', len(intersection_list))

# Convert list of lists to list of strings
intersection_list_2 = []

for i in intersection_list:
    intersection_list_2.append(i[0])

print('updated intersection list:', intersection_list_2)

Regex 2.2: Use text data to create start and end points

temp = intersection_list_2[0].split('at')
temp_2 = temp[]
temp_2 = temp[0] + '&' + temp_

# Instantiate df to hold output intersections
columns = ['city', 'start', 'end', 'tweet']
output_tweets_df = pd.DataFrame(index=range(len(city_tweets)), columns=columns)
output_tweets_df.fillna('', inplace=True)
output_tweets_df

output_tweets_df.loc[:, 'city'] = city
output_tweets_df

index = 0
for string in intersection_list_2:
    print(string)
    output_tweets_df.loc[index, 'tweet'] = city_tweets[index]
    # Pattern 1
    if 'between' in string:
        split_1 = string.split('between') # Split on 'between' first
        split_2 = split_1[1].split('and')
        start = split_1[0] + '&' + split_2[0]
        end = split_1[0] + '&' + split_2[1]
        output_tweets_df.loc[index, 'start'] = start
        output_tweets_df.loc[index, 'end'] = end
        print('start:', start)
        print('end:', end)
    # Pattern 2
    elif 'at' in string:
        split_1 = string.split('at') # Split on 'at' first
        split_2 = split_1[1].split(',')
        start = split_1[0] + '& ' + split_2[0]
        end = split_1[0] + '& ' + split_2[1]
        output_tweets_df.loc[index, 'start'] = start
        output_tweets_df.loc[index, 'end'] = end
        print('start:', start)
        print('end:', end)
    else:
        print('no matching patterns found for tweet at index:')
    index += 1
    print()

# View
output_tweets_df

# Final Clean up (maual)
output_tweets_df.loc[0, 'end'] = output_tweets_df.loc[0, 'end'].replace('  stopped traffic back to', '')

#### Final Output

output_tweets_df.drop_duplicates(inplace=True)
output_tweets_df

output_tweets_df.to_csv('../data/3-final/output_tweets', index=False)
