==> notebooks/1A_Historical_Tweets.ipynb <==
# "Historical" Tweets - from Twitter Search API using Tweepy Library
Preliminary steps before using Twitter's API:
1. Sign-up for a twitter account
2. Register a twitter developer account (requires email or phone number)
3. Create a developer app (I went with the name BlockedRoads)
4. Obtain your 'Access token' and 'access token secret' in the developer dashboard

Load Imports

import pandas as pd
import tweepy
import shutil
import jsonpickle
import json
import datetime
import config

from pandas.io.json import json_normalize

## Part 1: Prepare API credentials
**Step 1: Authenticate account and tokens through tweepy**
<br>Required credentials:
<br>`config.consumer_key` = 'your_consumer_key'
<br>`config.consumer_secret` = 'your_consumer_secret'
<br>`config.access_token` = 'your_access_token'
<br>`config.access_token_secret` = 'your_access_token_secret'

# authenticate account with tweepy
auth = tweepy.OAuthHandler(config.consumer_key, config.consumer_secret)
auth.set_access_token(config.access_key, config.access_secret)

# create API object to pull data from twitter - and pass in authentication code
api = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True)

Now we are free to make Twitter API calls

**Step 2: Verify API is working with your account**
- Run this cell to check if program is working. Output is your twitter name

user = api.me()
print (user.name)

**Step 3: Use geo_search API to get the `place_id` for a particular location**
- `place_id` is a unique id created by Twitter that is assigned to different neighborhoods/cities/countries
- In the query parameter, type the name of the city or country you want the place_id for
- granularity = neighborhood (default) , city , admin or country
    - EX: when query = 'Boston', the place_id generated is the same for neighborhood or city ( but empty for country)

#  Determine the city, country, or location you want
places = api.geo_search(query="Houston", 
                        granularity="city"
                                          )
place_id = places[0].id
print('place_id is: ',place_id)

***List of search queries made:***
<br>NOTE: there must be no space between place and place_ID

    
|searchQuery| place ID |granularity| file | # of tweets| extended tweet? |
|---|---|---|---|---|---|
|'place:96683cc9126741d1 road closed' | USA |country| ./PoGo_USA_Tutorial.json20190114_7_1_36.json | 142 | - |
| 'place:1c69a67ad480e1b1 road closed' | Houston | city/neigh | ./historical_20190114_11_59.json| 15|- |
| 'Houston road closed' | NA | NA | ./historical_20190114_12_2.json | 9|- |
|'place:96683cc9126741d1 road closed' | USA |city| ./historical_20190115_14_57.json | 210 | yes |
|'place:1c69a67ad480e1b1 road closed' | Houston |city| ./historical_20190117_11_49.json | 13 | no |


## Part 2: Download Archived Tweets with the Search API
**Step 1: Use tweepy library to download tweets from Search API based on our query search terms, and save tweets as a json file**
 - Include a place ID in the query if necessary
 - Other search params that are optional: since='2019-01-03',until='2019-01-11'
 
useful resource: http://www.dealingdata.net/2016/07/23/PoGo-Series-Tweepy/

searchQuery = 'place:1c69a67ad480e1b1 road closed'

tweetCount = 0

#Open a text file to save the tweets to
with open('../data/AT_historical/historical.json', 'w') as f:

    #Tell the Cursor method: we want to use Search API (api.search), and our query
    for tweet in tweepy.Cursor(api.search,q=searchQuery,
#                                tweet_mode='extended'    #comment out extended tweet if desired
                              ).items() :         

        #Verify the tweet has place info before writing 
        if tweet.place is not None:
            
            #Write the JSON format to the text file, and add one to the number of tweets we've collected
            f.write(jsonpickle.encode(tweet._json, unpicklable=False) + '\n')
            tweetCount += 1

    #Display how many tweets we have collected
    print("Downloaded {0} tweets".format(tweetCount))

**Step 2: Append a timestamp to the end of the json file name**

def file_conversion():
    #create a timestamp
    now = datetime.datetime.now()
    month = '0'+str(now.month)
    day = str(now.day)
    hour = str(now.hour)
    minute = str(now.minute)
    now_str = str(now.year)+month+day+'_'+hour+'_'+minute
    
    
    #replace the name of our file with a new timestamped filename
    dest = '../data/AT_historical/historical_' + now_str + ".json"
    shutil.move('../data/AT_historical/historical.json', dest)
    
    with open(dest, "r") as f:
        status = f.readlines()
        jsons = []
        for ind in status:
            jsons.append(json.loads(ind))
    return jsons

# Uncomment out the file_conversion function to activate the file name change
# file_conversion()

**Step 3: Add the newly created json file to the list below for easy access later**

# Previously created json files can be accessed from this list

# json_df = pd.read_json("../data/AT_historical/PoGo_USA_Tutorial.json20190114_7_1_36.json", lines = True)
# json_df = pd.read_json("../data/AT_historical/historical_20190114_11_59.json", lines = True)
# json_df = pd.read_json("../data/AT_historical/historical_20190114_12_2.json", lines = True)
# json_df = pd.read_json("../data/AT_historical/historical_20190115_14_57.json", lines = True)
json_df = pd.read_json("../data/AT_historical/historical_20190117_11_49.json", lines = True)

# Part 3: Explore Tweets from json file
### Option A: Tweets --> json file --> dataframe

# list of columns currently available in our dataframe
json_df.columns

**View all tweets from the `text` column** (for the first 7 tweets)

list = json_df.loc[:6, 'text']
for i in range(len(list)):
    print(i, list[i])

Explore attributes of the json file

def PrintMembers(obj):
    for attribute in dir(obj):
        #We don't want to show built in methods of the class
        if not attribute.startswith('__'):
            print(attribute)
PrintMembers(tweet)

**Shrink dataframe down to only the useful columns**

columns = ['coordinates', 'created_at', 'geo', 'place', 'text', 'user']
df = json_df.loc[:, columns]
df.head()

### Explore some features of the df

# see all the nested information within place (for the first two tweets)
list = json_df.loc[:1, 'place']
for i in range(len(list)):
    print(i, list[i])
    print("--------")

> The dataframe that we have now has many attributes nested within other attributes. If we want to access the corrdinates nested within `coordinates` or the city nested within `place` we will need to unpack these variables using a different method. Use the next section to convert nested values into their own columns.

## Option B. Tweets --> json --> parsed --> nested dataframe

# Read in json file
json_df = pd.read_json("../data/AT_historical/PoGo_USA_Tutorial.json20190114_7_1_36.json", lines = True)

# convert json file to list of dictionaries
tweets_data = []
notParsed = []
tweets_file = open("../data/AT_historical/PoGo_USA_Tutorial.json20190114_7_1_36.json","r")
for line in tweets_file:    
    if line.strip():    
        try:
            tweet=json.loads(line)
            tweets_data.append(tweet)
        except:
            notParsed.append(line)
            continue
print(len(tweets_data))
print('Could not parse: ', len(notParsed))

**Unpack the nested columns**

tweet_cols = ['coordinates', 'created_at', 
#                'full_text',
              'text','geo', 'id', 'place', 'user']

sample_tweets_dict = [{col:tweet[col] for col in tweet_cols } for tweet in tweets_data]

# This method looks for any instances of nested dictionaries (DOES NOT FLATTEN LISTS)
sample_tweets_df = pd.io.json.json_normalize(sample_tweets_dict)

sample_tweets_df.head(2)

sample_tweets_df.columns

# save the USA (or Houston) historical json file as df, COMMENT OUT
# df = sample_tweets_df
# df.to_csv("../data/AT_historical/USA_142tweets.csv", index = False)

Examine who the most common tweets are coming from

sample_tweets_df['user.screen_name'].value_counts().head()

So the Twitter user "TotalTraffic", specific to a certain city, is contributing the most tweets to situations with "road closed", and providing location place data for each tweet
