==> notebooks/1B_Stream_tweets.ipynb <==
# "Live" Tweets from Stream API using Tweepy Library

Preliminary steps before using Twitter's API:
1. Sign-up for a twitter account
2. Register a twitter developer account (requires email or phone number)
3. Create a developer app (I went with the name BlockedRoads)
4. Obtain your 'Access token' and 'access token secret' in the developer dashboard

## Part 1: Pull Live Tweets from Twitter using Tweepy and Streaming API - StreamListener

Some resources:
http://socialmedia-class.org/twittertutorial.html
<br>https://www.dataquest.io/blog/streaming-data-python/


**Step 1: In terminal run:** `! pip install tweepy`

**Step 2: Load imports**

import pandas as pd
import tweepy
import json
import datetime
from tweepy.streaming import StreamListener
import shutil
import config

**Step 2: Authenticate Twitter Credentials through Tweepy**

# authenticate account with tweepy
auth = tweepy.OAuthHandler(config.consumer_key, config.consumer_secret)
auth.set_access_token(config.access_key, config.access_secret)

# create API object to pull data from twitter - and pass in code!
api = tweepy.API(auth, wait_on_rate_limit=True, wait_on_rate_limit_notify=True)

# now we should be free to make twitter api calls!

**Step 3: Verify API is working with your account**

# check if program is working - output is your twitter name
user = api.me()
print (user.name)

# To send a Tweet from your account (assuming you allowed your app to have read/write permissions):
# api.update_status(status = "Hey, I'm tweeting with Tweepy!")

**Step 4: Instantiate the StreamListener class and start pulling tweets**
- Specify how tweets are downloaded inside the `on_status` or `on_data` class method
- `stream.filter` starts the download. Specify your search words with the `track` parameter
- To stop the live stream, **kernel interrupt the codeblock**
- Continue to Part 2 to save the tweets into a json file

class StreamListener(tweepy.StreamListener):            
    def on_data(self, status):   
        try:
            with open('../data/AT_stream/scrape.json', 'a') as f:
                f.write(status)
        except:
            print("Could not open file log")
        
         
    def on_error(self, status_code):
        if status_code == 420:
            return False

        
stream_listener = StreamListener()
stream = tweepy.Stream(auth=auth, listener= stream_listener)
# starts the stream
stream.filter(track=[ "road closed", "roads closed", "road blocked", "street blocked", "roadclosed"])

## Part 2: Save Tweet to a JSON file

def file_conversion():
    #create a timestamp
    now = datetime.datetime.now()
    month = '0'+str(now.month)
    day = str(now.day)
    hour = str(now.hour)
    minute = str(now.minute)
    now_str = str(now.year)+month+day+'_'+hour+'_'+minute
    
    
    #replace the name of our file with a new timestamped filename
    dest = '../data/AT_stream/scrape_' + now_str + ".json"
    shutil.move('../data/AT_stream/scrape.json', dest)
    
    with open(dest, "r") as f:
        status = f.readlines()
        jsons = []
        for ind in status:
            jsons.append(json.loads(ind))
    return jsons

# Uncomment out the file_conversion function to activate the file name change
# file_conversion()

**Table of most recent stream downloads**

|Stream| params| json file| # of tweets |search terms|
|---| --- | --- | --- | ---|
|1| tweet_mode='extended'| './scrape_20190115_11_28.json' | 1| track1|
|2| none?| './scrape_20190116_12_15.json'  | 4| track1|

track1 =[ "road closed", "roads closed", "road blocked", "street blocked", "roadclosed"]


# Previously created json files
# json_df = pd.read_json("../data/AT_stream/scrape_20190116_11_28.json", lines = True)
json_df = pd.read_json("../data/AT_stream/scrape_20190116_12_15.json", lines = True)

print(json_df.shape)
json_df.head()

## Part 3: Unpack the nested dictionary into separate columns

tweets_data = []
notParsed = []
tweets_file = open("../data/AT_stream/scrape_20190116_12_15.json","r")
for line in tweets_file:    
    if line.strip():    
        try:
            tweet=json.loads(line)
            tweets_data.append(tweet)
        except:
            notParsed.append(line)
            continue
print(len(tweets_data))
print('Could not parse: ', len(notParsed))

tweet_cols = ['coordinates', 'created_at', 
#                'full_text',
              'text','geo', 'id', 'place', 'user']


sample_tweets_dict = [{col:tweet[col] for col in tweet_cols } for tweet in tweets_data]

# This method looks for any instances of nested dictionaries (DOES NOT FLATTEN LISTS)
sample_tweets_df = pd.io.json.json_normalize(sample_tweets_dict)
sample_tweets_df

Let's look at the tweets we collected during the live stream

list = json_df.loc[:, 'text']
# print(list)
for i in range(len(list)):
    print(i, list[i])
    print("--------")
