==> notebooks/1.5-beb-gmaps-tool-learning.ipynb <==
# Dual-Purpose Notebook

The first part of this notebook consists of space dedicated to testing and familiarizing ourselves with the gmaps functionality, not all of which was used in our final product.

The second tests for, then defines the function used to render our final maps.

import gmaps
import gmaps.datasets
import python_scripts.config
import pandas as pd
import numpy as np
import time
gmaps.configure(api_key= python_scripts.config.gmaps_key)


## Areas of Interest

The first 3 figures are the basic maps of some areas of interest. For our presentation we decided to stay close to home and the rest of the project will deal primparily with the Seattle area.

figure_layout = {
    'width': '700px',
    'height': '400px',
    'padding': '1px',
    'margin': '0 auto 0 auto'
}

houston_coordinates = (29.7604, -95.3698)
fig = gmaps.figure(center=houston_coordinates, zoom_level=10,layout=figure_layout)
#fig.add_layer(gmaps.traffic_layer())
fig

figure_layout = {
    'width': '700px',
    'height': '400px',
    'padding': '1px',
    'margin': '0 auto 0 auto'
}

boston_coordinates = (42.3601, -71.0589)
fig = gmaps.figure(center=boston_coordinates, zoom_level=12,layout=figure_layout)
#fig.add_layer(gmaps.traffic_layer())
fig

figure_layout = {
    'width': '700px',
    'height': '400px',
    'padding': '1px',
    'margin': '0 auto 0 auto'
}

seattle_coordinates = (47.6062, -122.3321)
fig = gmaps.figure(center=seattle_coordinates, zoom_level=10,layout=figure_layout)
#fig.add_layer(gmaps.traffic_layer())
fig

## Code Examples

These are examples of code provided by the gmaps website for demonstration purposes. I'll be drawing from them for reference.

fig = gmaps.figure(center=(51.5, 0.1), zoom_level=9)

# Features to draw on the map
gmt_meridian = gmaps.Line(
    start=(52.0, 0.0),
    end=(50.0, 0.0),
    stroke_weight=3.0
)
greenwich = gmaps.Marker((51.3, 0.0), info_box_content='Greenwich')

drawing = gmaps.drawing_layer(features=[greenwich, gmt_meridian])
fig.add_layer(drawing)
fig

# This functionality was not used in our present build

london_congestion_zone_path = gmaps.datasets.load_dataset('london_congestion_zone')

fig = gmaps.figure(center=(51.5, -0.1), zoom_level=12)
london_congestion_zone_polygon = gmaps.Polygon(
    london_congestion_zone_path,
    stroke_color='blue',
    fill_color='blue'
)
drawing = gmaps.drawing_layer(
    features=[london_congestion_zone_polygon],
    show_controls=False
)
fig.add_layer(drawing)
fig

# This functionality was not used in our present build

london_congestion_zone_path[:2]

geneva = (46.2, 6.1)
montreux = (46.4, 6.9)
zurich = (47.4, 8.5)

fig = gmaps.figure()
geneva2zurich = gmaps.directions_layer(geneva, zurich)
fig.add_layer(geneva2zurich)
fig

fig = gmaps.figure()
geneva2zurich_via_montreux = gmaps.directions_layer(
        geneva, zurich, waypoints=[montreux],
        travel_mode='DRIVING')
fig.add_layer(geneva2zurich_via_montreux)
fig

fig = gmaps.figure()
geneva2zurich = gmaps.directions_layer(
    geneva, zurich, show_markers=False,
    stroke_color='red', stroke_weight=3.0, stroke_opacity=1.0)
fig.add_layer(geneva2zurich)
fig

nuclear_power_plants = [
    {'name': 'Atucha', 'location': (-34.0, -59.167), 'active_reactors': 1},
    {'name': 'Embalse', 'location': (-32.2333, -64.4333), 'active_reactors': 1},
    {'name': 'Armenia', 'location': (40.167, 44.133), 'active_reactors': 1},
]

fc = [{'text':'text 1'},{'text':'text 2'},{'text':'text 3'}]

plant_locations = [plant['location'] for plant in nuclear_power_plants]
info_box_template = """
<dl>
<dt>text</dt><dd>{text}</dd>
</dl>
"""
plant_info = [info_box_template.format(**road) for road in fc]

marker_layer = gmaps.marker_layer(plant_locations, info_box_content=plant_info)
fig = gmaps.figure()
fig.add_layer(marker_layer)
fig

## Fake Location Data

A table of data in the format I want for the final product, for use in testing

data = {'start':[[47.608222,-122.334897],[47.607477, -122.334210],[47.609730, -122.337820]],
'end':[[47.610146, -122.336646],[47.606116, -122.337564],[47.610663, -122.335567]],
'text':['Fallen Power Line',None,'Collision']}
fake_closures = pd.DataFrame(data)
fake_closures.head()

## Funcionality

Running code to test the render_closures programming

# checking my iteration logic...
for index,point in fake_closures.iterrows():
    print(index)
    print(point['start'])

list(fake_closures['text'])

# testing some simple road closures
# first establish the figure layout:
figure_layout = {
    'width': '700px',
    'height': '400px',
    'padding': '1px',
    'margin': '0 auto 0 auto'
}

# instantiate the figure with said layout
fig = gmaps.figure(layout=figure_layout)

# establish list for map markers
marker_coords = []
marker_labels = []
# iterate over the index of rows
for index,_ in fake_closures.iterrows():
    # draw the closure, using directions instead of lines to ensure it follows the road
    # accurately
    close = gmaps.directions_layer(start=fake_closures.iloc[index][0],
                                   end=fake_closures.iloc[index][1],
                                   travel_mode='WALKING',show_markers=False,
                                   stroke_color='Red', stroke_opacity=0.8)
    # add the closure to the map
    fig.add_layer(close)
    # add to our label text column
    if fake_closures.iloc[index][2] is not None:
        # append label info (I'm not really happy with how these end up looking, but
        # we're on a time crunch. Function over form, I say!)
        marker_coords.append(fake_closures.iloc[index][0])
        marker_labels.append(fake_closures.iloc[index][2])

# if marker_labels isn't empty
    if marker_labels != []:
        # establish labeled markers
        marker_layer = gmaps.marker_layer(locations=marker_coords,
                                  label=marker_labels)
        # and add them to the map
        fig.add_layer(marker_layer)
# add that layer of labels
fig.add_layer(marker_layer)
# render
fig

## Define Function

Using the above code, we now wrap the process into a function for use in other notebooks.

def render_closures(data):
    # import modules for the python script:
    import pandas as pd
    import gmaps
    import time

    # layout options: draw a comfortable-sized map
    figure_layout = {
    'width': '700px',
    'height': '400px',
    'padding': '1px',
    'margin': '0 auto 0 auto'
    }
    # initialize map
    fig = gmaps.figure(layout=figure_layout)

    # establish list for map markers
    marker_coords = []
    marker_labels = []

    # draw road closures
    for index,_ in data.iterrows():
        # The parameters are mostly self-explanatory, but I will note that this uses
        # walking directions in case start and end points are contrary to one-way streets
        close = gmaps.directions_layer(start=data.iloc[index][0],end=data.iloc[index][1],
                                       travel_mode='DRIVING',show_markers=False,
                                      stroke_color='Red', stroke_opacity=0.8)
        # add this closure's layer
        fig.add_layer(close)
        # check for label text:
        if data.iloc[index][2] is not None:
            # append label info (I'm not really happy with how these end up looking, but
            # we're on a time crunch. Function over form, I say!)
            marker_coords.append(data.iloc[index][0])
            marker_labels.append(data.iloc[index][2])

        # Google Maps API limits to 50 requests per second
        time.sleep(0.3)

    # if marker_labels isn't empty
    if marker_labels != []:
        # establish labeled markers
        marker_layer = gmaps.marker_layer(locations=marker_coords,
                                  label=marker_labels)
        # and add them to the map
        fig.add_layer(marker_layer)
    return fig


render_closures(fake_closures)

## Export Example DataFrame

For test purposes in other dataframes

fake_closures.to_csv('./coords-df.csv',index=False)
