==> notebooks/2.0-beb-geocoding-intersections.ipynb <==
# Geocoding Intersections

As an intermediate step in processing social media text, I will use geopy and gmaps to convert strings of road intersections into latitude and longitude coordinates for use in our render_closures function.

import pandas as pd
import geopy
import gmaps as gmaps
from python_scripts.render_closures import render_closures
import python_scripts.config as config
gmaps.configure(api_key= config.gmaps_key)

## Fake Data

To test populating our map.

fake_closures = {
    'start':[
        'rainier ave s & s edmunds st',
        's oregon st & 38th ave s',
        's genesee st & 42nd ave s'
    ],
    'end':[
        '35th ave s & s edmunds st',
        's oregon st & rainier ave s',
        's alaska st & 42nd ave s'
    ],
    'city':[
        'seattle',
        'seattle',
        'seattle'
    ],
    'text':['Fallen Power Line','Construction','Collision']
}
fake = pd.DataFrame(fake_closures)
fake.head()

## Functionality

Exploring and establishing the code for our function: take the dataframe in and return latitude and longitude start and end points, for use in the render_closures function

# instantiate a google maps geocoder
geolocator = geopy.geocoders.GoogleV3(api_key=config.gmaps_key)
# extract the latitude and longitude
location = geolocator.geocode(fake.start[0])
# print it to check
print((location.latitude, location.longitude))


That works so far, but that may only be because my IP address is *within* the city in question. Let's try adding in the city, just to be sure.

# instantiate a google maps geocoder
geolocator = geopy.geocoders.GoogleV3(api_key=config.gmaps_key)
# extract the latitude and longitude
location = geolocator.geocode(fake.start[0] +',' + fake.city[0])
# print it to check
print(location)

## Building a DataFrame

That seems to work, so now we can iterate through our intersection-based DataFrame and output a shiny new lat/long-based DataFrame and write a function to do just that.

# instantiate a google maps geocoder
geolocator = geopy.geocoders.GoogleV3(api_key=config.gmaps_key)
# build lists to put into our new DataFrame
starts = []
ends = []
texts = []

# iterate over rows
for index,point in fake.iterrows():
    # geocode the start and end points
    new_start = geolocator.geocode(fake.start[index] +',' + fake.city[index])
    new_end = geolocator.geocode(fake.end[index] +',' + fake.city[index])
    # put their latitudes and longitudes into our lists
    starts.append([new_start.latitude,new_start.longitude])
    ends.append([new_end.latitude,new_end.longitude])
    texts.append(fake.text[index])

# put all that together!
new_fake = {'start':starts,'end':ends,'text':texts}
new_df = pd.DataFrame(new_fake)
new_df

fake

render_closures(new_df)

def intersect_to_coords(data,exdata=None):
    import pandas as pd
    import geopy
    # instantiate a google maps geocoder
    geolocator = geopy.geocoders.GoogleV3(api_key=config.gmaps_key)
    # build lists to put into our new DataFrame
    starts = []
    ends = []
    texts = []

    # iterate over rows
    for index,point in data.iterrows():
        # geocode the start and end points
        new_start = geolocator.geocode(data.start[index] +',' + data.city[index])
        new_end = geolocator.geocode(data.end[index] +',' + data.city[index])
        # put their latitudes and longitudes into our lists
        starts.append([new_start.latitude,new_start.longitude])
        ends.append([new_end.latitude,new_end.longitude])
        texts.append(data.text[index])

    # put all that together!
    new_dict = {'start':starts,'end':ends,'text':texts}
    new_df = pd.DataFrame(new_fake)
    
    if exdata is not None:
        new_df = pd.concat([new_df,exdata],ignore_index=True)
        tuple(new_df['start'])
        tuple(new_df['end'])
    return new_df

df3 = intersect_to_coords(fake)
render_closures(df3)

from ast import literal_eval

coords = pd.read_csv('./coords-df.csv',
                     converters={"start": literal_eval,'end':literal_eval})
coords

df = intersect_to_coords(fake,coords)
df

df.start[3]
# for index,_ in df.iterrows():
#     print(df.start[index])
#     print(type(df.start[index]))
#     print('---')
#     print(df.end[index])
#     print(type(df.end[index]))
#     print('---')
#     print(df.text[index])
#     print(type(df.text[index]))
#     print('*****')

render_closures(df)

## Mission accomplished!

I manually saved the above function into another file in python_scripts for future use
