==> notebooks/3.1-beb-here-mapping.ipynb <==
import pandas as pd
import geopy
import gmaps as gmaps
from python_scripts.render_closures import render_closures
import python_scripts.config as config
from ast import literal_eval
gmaps.configure(api_key= config.gmaps_key)

# Rendering Here.com Closures

The purpose of this notebook is to import, format, and render data from [here.com](https://www.here.com/) into our `render_closures` function. Data was collected and exported to csv in [this](1.0-bk-heremaps-data-extraction.ipynb) notebook.

## Read Data

Here I read in our example of here.com road closures and format it to a usable state.

here = pd.read_csv('../data/2-interim/here_maps_output_2019-01-15-11PM')
here.info()

features = ['start_lat_long','end_lat_long']
df = pd.DataFrame(columns=['start','end','text'])
df['start'] = [literal_eval(here['start_lat_long'][i]) for i in range(len(here))]
df['end'] = [literal_eval(here['end_lat_long'][i]) for i in range(len(here))]
df['text'] = None
df.head(10)

## Rendering the Data

Our data is now properly formatted for us in the render_closures function, but our gmaps rendering is limited by Google's API restrictions to only display 10 road closures.

test_df = df.head(10)

render_closures(test_df)

## Function

Quickly turning the process of formatting the here.com data into a usable format into a function.

def format_here_data(data):
    # import necessary modules
    import pandas as pd
    from ast import literal_eval
 
    # establish the features from data that we'll keep
    features = ['start_lat_long','end_lat_long']
    # establihs the dataframe
    df = pd.DataFrame(columns=['start','end','text'])
    # read in the coordinates, evaluating columns because read_csv reads lists and tuples
    # as strings
    df['start'] = [literal_eval(data['start_lat_long'][i]) for i in range(len(data))]
    df['end'] = [literal_eval(data['end_lat_long'][i]) for i in range(len(data))]
    df['text'] = None
    
    return df

test_fun = format_here_data(here)
test_fun.head()

render_closures(test_fun.head(10))
