==> notebooks/4.1-beb-final-data-mapping.ipynb <==
import pandas as pd
import gmaps
from ast import literal_eval

from python_scripts.render_closures import render_closures
from python_scripts.intersect_to_coords import intersect_to_coords
from python_scripts.format_here_data import format_here_data
import python_scripts.config as config

gmaps.configure(api_key= config.gmaps_key)

# The "Finished" Product

In this notebook I will render a map of road closures including data from bother here.com and Twitter's API, to include in our presentation and to show the workflow of the map making process.

## Read and Format Data

# read in the data
here_data = pd.read_csv('../data/2-interim/here_maps_output_Houston, TX_2019-01-17-12PM')
tweets = pd.read_csv('../data/3-final/output_tweets')

# tweets format wasn't quite right. Let's try this:
tweets.rename(index=str, columns={"tweet": "text"},inplace=True)
tweets

# format the here.com data
here_data = format_here_data(here_data)
here_data

# geocode the Twitter intersections, append the closures to here_data
# Google Maps API doesn't allow more than 10 closures per map
all_closures = intersect_to_coords(tweets,here_data)[:11]

# I'll manually alter the two labels we have. Future versions of the project could use a
# model to extract the relevant label data
all_closures.text[:2] = 'road construction'
all_closures

all_closures.info()

# One row's start and end points were the same, which throws errors
all_closures.drop([0,8],inplace=True)
all_closures.reset_index(drop=True,inplace=True)
all_closures

## Render Map

# Now we render the road closures using Google Maps
render_closures(all_closures)

## Export Final DataFrame

For thoroughness's sake.

all_closures.to_csv('../data/3-final/final-closure-df.csv')
